Bug#33394701: crash at JOIN::optimize_keyuse()
When setting up multiple equalities, an outer reference
is added when should not be. The problematic field is a
view reference referring to another view reference.
Currently we check only "depended_from" for the view
reference to determine if its an outer reference which
does not suffice in this case. We also need to check if
the underlying reference is an outer reference. So the
fix is to make a call to "is_outer_reference()" rather
than checking only "depended_from".

Change-Id: I50ab39ca448fe48e7b9cdb5b150c5000de2ec498