Bug#33671403 WL14864 Setting --DEFAULT.* does not always propagate command line overrides
In the scenario where there is no configuration file used (MCM, not possible
with MySQL Router) the initial default section options are not overwritten with
--DEFAULT.xxxx parameter.
The problem is that the Config constructor overload that takes initial defaults
as a parameter fails to apply the overwrites and since there is no config file
no other method is called later that would force applying them.

This patch adds apply_overwrites call to the constructor mentioned above.

RB: 27473