Bug#33611112: Reverse index scan is not sorted on implicit InnoDB primary key
The hypergraph optimizer returned rows in the wrong order when using a
reverse index range scan on a secondary index in order to get rows
sorted on the primary key in descending order.

Reverse index range scans need to be told up front that they should
return rows with equal keys in descending primary key order. This is
so that they can use regular forward order scans (which are presumably
faster than reverse scans) for equal ranges.

The hypergraph optimizer did not set the flag to use the extended key
parts when it needed them. This patch makes it set it on reverse index
range scans if returning otherwise equal rows in descending primary
key order would satisfy more interesting orderings than if the rows
are ordered only by the user-defined key parts.

Change-Id: I8220c14041675740ec2ffd00af45956660705453