Bug#33491183: Wrong query result using select Union All Bug#33503357: comparing a `bigint` value with an identical `bigint unsigned` value fails Bug#33634109: hypergraph: Unstable results from TPC-H Q2 Bug#33674441: Wrong result with outer join and streaming aggregation
Aggregated queries that performed an index lookup (eq_ref) could in
some cases return wrong results. This is a regression after
bug#31790217 (REWRITE STREAMING AGGREGATION).

The bug is only seen if the execution plan used streaming aggregation
instead of aggregation in temporary tables. (Since streaming
aggregation currently is the only aggregation strategy supported by
the hypergraph optimizer, the bug is seen more frequently with the
hypergraph optimizer. But the bug also affects the old optimizer.)

The issue is that AggregateIterator and EQRefIterator touch the record
buffer (table->record[0]) of the same TABLE object. EQRefIterator has
a mechanism for caching the last key that it was called with, in order
to avoid a new index lookup if it is called again with the same key.
It can do this on the assumption that the contents of the record
buffer still point to the row with the requested key. However,
AggregateIterator's modifying of the same record buffer breaks this
assumption. Specifically, when AggregateIterator reads a row and
detects that it has come to the next group, it modifies the record
buffers of the underlying scans to point into an arbitrary row in the
previous group. This means that the EQRefIterator below it cannot rely
on its record buffer to point to the last row it returned, and it
cannot reuse those values even if the key is the same as in the
previous call.

This patch fixes the bug in the hypergraph optimizer by adding a
STREAM access path just below the AGGREGATE access path, if there is
an EQ_REF. This makes the AggregateIterator work on a different TABLE
object than the EQRefIterator, so that it doesn't interfere with
EQRefIterator's caching.

A similar fix could have been applied to the old optimizer. But adding
intermediate temporary tables is harder in the old optimizer, so the
patch simply disables EQ_REF caching below AGGREGATE nodes in the old
optimizer for now.

Change-Id: Idd7bda5dfca24052bfc593f036a95e13f9bf1644