Bug #33045204 EQUI-HEIGHT HISTOGRAM ALWAYS ESTIMATES NON-INTEGRAL EQUALITY PREDICATE TO 1/ROWS [patch 2/2, efficiency fix]
The two types of histograms supported by MySQL, singleton and
equi-height, store their collection of buckets in a std::map and
std::set, respectively.

Part two of this patch changes both histogram types to store buckets in
a dynamic array instead. Storing buckets in a dynamic array reduces
space overhead and speeds up selectivity estimation due to less
indirection when performing binary search on the buckets.

Because both types of histograms are constructed by inserting buckets
in sorted order, and buckets are serialized to JSON in sorted order,
there is no additional work involved in switching to a dynamic array.

As an example of the gain in performance, the running time of the tests
in histogram_selectivity-t was reduced locally from 2.7 seconds to 1.7
seconds.

This patch also changes singleton and equi-height histograms to be
constructed through factory methods. This is done in order to ensure
that we catch allocation errors during construction and avoid returning
histograms that have not been initialized correctly.

Change-Id: If04928771e90d7651d8dd47ff0606700468fc0b0