Bug#33611545 - regression: 8.0.27+ crash in heap_rkey (MEMORY) or number_of_indexes (TempTable)
In case of a MATERIALIZE access path the non-basic(sort,filter
etc) iterators are moved from the table_path to above the path
for better EXPLAIN and simpler structure.

An INDEX_RANGE_SCAN access path is also not considered a basic
iterator and a traversal is attempted to find the underlying
basic iterator. In cases where the index reads are enough and
there is no other access the behaviour becomes undefined and may
cause server exit.

Fix:
Consider INDEX_RANGE_SCAN to be a basic iterator and need not be
moved from the table access path

Change-Id: Iffde5108be47ee633b09ad5ee613bd89c7d51204