Bug#33727629 Misc fixes for ndb_file on Windows
ndb_file::OFF_T_MAX was declared unsigned long but intended to be a
64-bit value, on Windows long is 32 bit - changed to use Uint64.

In ndb_file::extend on Windows failure of SetFileValidData is ignored,
but system error was not cleared which could confuse when other
functions failed later.

And also in ndb_file::create on Windows failure of CloseHandle is
ignored without clearing system error.

Added clearing of system error.

In ndb_file::open in Windows FILE_FLAG_RANDOM_ACCESS was always used even if
file was accessed sequentially - removed use of that flag.

Change-Id: Ia90c1177e38d0bf3a415280988b33c8ef5999235