Bug #33635120 Rename character set names from 'utf8' to 'utf8mb3'
'utf8' is a deprecated alias for 'utf8mb3'
However, in our implementation, it was the other way around:
character sets were named 'utf8' and 'utf8mb3' was the alias.

A long series of earlier patches have changed the output from the server:
Bug #nnnnnnnn CHANGE 'UTF8' TO 'UTF8MB3' for .......
so that we always output 'utf8mb3' rather than 'utf8'.

This patch does the actual internal renaming of the character sets.

For external character sets, we do the renaming in XML parser.

Change-Id: Iae5fa51b6c1d47d0f14d373f9f668e38895dcacb