Bug#33583435 - Result mismatch seen with view Bug#33583479 - Mysql server crash - Assertion `!value.isnull' failed|handler/protocol_adapter.cc
Query
===============
CREATE VIEW v0 AS
SELECT t00.a, t01.a AS b, IFNULL(t01.a, 666) AS c
FROM t00 LEFT JOIN t01 USING(a);
SELECT * FROM v0 WHERE c >= 1;

On Heatwave:
============
+------+------+---+
| a    | b    | c |
+------+------+---+
|    1 |    1 | 1 |
+------+------+---+
On MySQL:
=========
+------+------+-----+
| a    | b    | c   |
+------+------+-----+
|    1 |    1 |   1 |
|    2 | NULL | 666 |
+------+------+-----+

Previously : As there was a REF_ITEM around c we were computing it during the table scan of t01.

Solution :
===================
let item = VIEW_REF(IFNULL(t01.a, 666))
the first_inner_table of item is nullptr, hence we do not pushdown the expression.

first_inner_table tells us that if an item belongs to a view
that is an inner table of an outer join, then this field points to the first leaf table of the view,
otherwise NULL.

Change-Id: Icd61fbad374578c17ef6ca2f76512c4562167c62