Bug #33712370 MySQL 8.0.23 SLAVE HUNG AFTER SEMAPHORE WAIT
An "ABA"-like race condition was possible in `Batch_segment::write_complete()` function - if the thread got scheduled out in between its two instructions:
```
const auto n = m_written.fetch_add(1, std::memory_order_relaxed);
return n + 1 == m_batch_size.load(std::memory_order_relaxed);
```
then it could happen that another thread has already completed the batch segment, called `reset()` and `enqueue()`d it, the segment got `dequeue()`d again, and used for batch with size exactly matching `n+1`.
This would lead to `write_complete()` erroneusly reporting that the batch has completed, and subsequently could lead to NPE or infinite wait.

One possible fix would be to change the order of the two operations.

One can also view this mechanisms as a variant of reference counting, where we consider the segment no longer needed when `m_written` reaches `m_batch_size`.
This patch reimplements this using a single atomic for refcouning: `m_uncompleted` which is set to match `m_batch_size` when starting the batch, and decremented whenever a page write is completed.
This simpler aproach makes it impossible to mess up the order of the two operations, as there's only one operation now.

The `m_batch_size` field is now only used for statistics, and no longer needs to be atomic (see proof in doxygen).

The `m_dblwr_id` field was marked as `private` as it probably always should given we only access it via getter and setter. I was considering changing it to atomic, so that `set_dblwr_batch_id()` in thread requesting the IO write could synchronize-with `get_dblwr_batch_id()` in IO completer, and thus make all of the fields `s_segments[batch_id]` available for reading, but then I've realized, that if we can not assume that IO completion happens-after IO request, then we have bigger troubles to worry about than that. We can assume this happens-before relation not only because it makes intuitive sense that IO can't finish before it started, but also because our own code synchronizes access to the slots array (in particular: to pass `m1` and `m2` meta-data from write requestor to completer).

All accesses to `m_uncompleted` to are `std::memory_order_relaxed` because as stated above we rely on happens-before relation between writer requestor and completer, and the code before the fix seemed to be avoiding synchronization wherever possible and I don't want to introduce a slowdown.

RB:27537
Reviewed by: Marcin Babij <marcin.babij@oracle.com>

Change-Id: Ie56bbe2da8c3ae84d10df72326659d2c95336d80