Bug#33619350: Hypergraph: Assertion `it != m_access_paths.end()' failed.
Follow-up patch for a similar problem as the one fixed in the original
patch for this bug. The hypergraph doesn't return a full plan for a
semi-join, because it rejects too many plans due to what it thinks is
a redundant join predicate.

The code that detects redundant join predicates and rejects plans that
use them, is deliberately too aggressive and could have false
positives. CostingReceiver::RedundantThroughSargable() mentions this:

// Note that there are more complicated situations, e.g. if t2 instead
// had t2.x = t4.x in the example above, where we could reject non-redundant
// join orderings. However, in nearly all such cases,
// DisallowParameterizedJoinPath() would reject them anyway, and it is not
// an issue for successfully planning the query, as there would always exist
// a non-parameterized path that we could use instead.

The last part, which says "there would always exist a
non-parameterized path that we could use instead", is not quite
accurate. It is true that a non-parameterized alternative is always
kept. But that assumes that the paths we are looking at actually are
parameterized. In the failing query, they are not parameterized. Both
left_path and right_path contain parameterized sub-paths (REF and
EQ_REF paths), but left_path and right_path themselves are not
parameterized, because they are nested-loop joins which already have
the required parameter tables on the outer side. Since we don't have
other non-parameterized paths to fall back on when we do this
rejection, we end up rejecting all alternatives we have for joining
the two sides of the semi-join, and no plan is returned.

This patch makes the rejection slightly less aggressive, to ensure we
will still find a valid plan in case we have a join predicate falsely
detected as redundant. It now only rejects the plan if the inner side
is parameterized on something on the outer side. Since this implies
that at least one of the paths is parameterized, and
ProposeAccessPath() always keeps a non-parameterized alternative, we
know that the comment above is correct, and we will be able to find
another valid plan using a non-parameterized alternative.

Change-Id: I673c305dab8fe05cb675cc679c12a45599a9b4cb