Bug#33457946: main.index_merge_innodb fails sporadically on pb2 daily trunk
After the commit titled
Bug #33037007: PRELIMINARY FIXES FOR WL #14488 [intersection num_rows, noclose]
both main.index_merge_myisam and main.index_merge_innodb started
failing regularly because of differences in the "filtered" column of
EXPLAIN output in some statements.

The underlying instability was not introduced by that commit. It only
removed a bug that hid the instability by always calculating
"filtered" to 100.00 for these queries.

The InnoDB variant of the test already masks out row estimates and
filter estimates for a lot of EXPLAIN statements, with comments about
known instabilities in the estimates. The patch adds similar masking
to some more statements.

The MyISAM variant of the test only fails on 32-bit platforms. During
an INSERT .. SELECT statement, the 64-bit binaries and the 32-bit
binaries choose to perform one of the index page splits at a slightly
different time, and thereby making the final structure of the B-tree
diverge a little. This results in ha_myisam::records_in_range()
returning a different estimate in two subsequent EXPLAIN statements
that are using the index.

The patch masks out this difference by hiding the "filtered" column in
the two affected EXPLAIN statements.

Change-Id: Ib45feb595ef30522c72cd650243b5ab6a6d79257