Bug#33165726 Fix freed memory access in performance schema table
Based on a contribution by Facebook.

Problem:

When executing
  TRUNCATE TABLE performance_schema.status_by_thread;
some freed memory can be used,
when inspecting the status variable of a session
which is concurrently disconnecting.

Root cause:

The implementation of this truncate statement,
reset_status_by_thread(),
performs a full scan on all PFS_thread instances,
as found in global_thread_container.

This scan is safe.

Then, for each PFS_thread object found,
the code follows a pointer to the associated THD object,
in PFS_thread::m_thd.

This is unsafe.

If the inspected session concurrently disconnects,
the THD object will be destroyed,
and the thread executing the truncate statement
will use a stale pointer.

Fix:

The implementation of
  TRUNCATE TABLE performance_schema.status_by_thread;
has been changed to not iterate on PFS_thread objects
in the performance schema space,
but iterate on THD objects in the SQL layer instead.

This iteration is done under the protection of
  Global_THD_manager::do_for_all_thd()
which guarantees that THD objects are safe to inspect
during the execution of this scan.

For each THD object found in the SQL layer,
the existing api
  PSI_THREAD_CALL(aggregate_thread_status)(thread);
is used to aggregate the status of each session.

This calls the same api as before:
  aggregate_thread_status(thread, account, user, host)
but this time with the guarantee that
- the PFS_thread object
- the related THD object
are safe to inspect.

Manual testing:

Used valgrind to expose the race conditions.

Reproduced the issue in valgrind without the fix.

Verified the issue in valgrind is resolved with the fix.

Change-Id: I75ced1f9b79679c76c9d5b67d30b9fce922cf129