Implemented PXB-1818 - Provide a new flag to improve performance by auto-tuning the value for --use-memory
https://jira.percona.com/browse/PXB-1818

Implementation:
For this implementation we keep resizing the buffer pool at prepare as
the parsed redo log records fill the buffer pool until we have enough
memory to hold all parsed records or we exceed the limit percentage of
free memory we can consume. A resize means we double the buffer pool
size. On each resize we also double the frames available to load
database pages.

Added a new flag to control the percentage of free memory we can
consume. It's important to note that we check the available
memory each time we are about to resize the buffer pool. To be on the
safe side, we don't consider the amount of memory we already hold on
buffer pool. For example:
If we have a BP of 1G we will go to 2G, if we selected
--use-free-memory-pct=50 to not pass 50% of free memory, this will
require at least 4G free of memory for this resize, not taking into
consideration that we already hold 1G.