Bug#33714885 - LOAD DATA INFILE fails with ERROR 1054 : Unknown column XXXX in 'where clause'
The field of the table being loaded into if used in the where clause
of a SET column, is reported as a unknown.

The outer context of the current query block is not set and trying to
resolve the name of a column of the outer select fails.

The outer context is set only if the parsing_context of a query bock
is non-NONE. In case of the LOAD DATA statement the parsing context is
not being set, hence the error.

Fix:
Set the parsing_context for the LOAD statement to be similar to a
statement with UPDATE_VALUE.

Change-Id: I898c6a84c443cf0a2a7efc3c46191e7bd709c5b7