Bug#33334928: SELECT COUNT(*) using multi-valued index reports wrong number of rows.
wl#8955 points out that while accessing the rows using
multi-value index there is a possibility of fetching the
same row multiple times. To avoid duplication of records the
unique record filter was enabled only for the range access,
and ref access was ignored assuming it will always do a
single key look-up and returns no-duplicate row.

However, while using ref or ref_or_null access we can
encounter the duplicate rows and thus the count(*) is
greater when the mv-index is used.

The fix is to enable unique record filter on the ref and
ref_or_null access paths.

Change-Id: I8f1458a585729cde61a2b91905d91af5d6d3af32