Bug#33743403: ASSERT: `false' in in() at join_optimizer/join_optimizer.cc
An assertion was raised to indicate that two plans for the same join
had different estimates for the number of output rows. This happened
when a query contained a predicate that had to be evaluated after the
join (a WHERE predicate that could not be evaluated until after an
outer join), and the predicate contained a materializable subquery.

The inconsistency was caused by reuse of AccessPath objects so that
the selectivity of the predicate was applied twice. Both
ProposeHashJoin() and ProposeNestedLoopJoin() would first add the
predicate to the join path without materialization, and then again
with materialization. And the selectivity of the predicate is applied
to the row estimate of the join path both times. This means the path
without materialization gets the selectivity applied once (which is
correct), whereas the path with materialization gets it applied twice.

Fixed by making a fresh copy of the join path before applying the
predicates, and applying the predicates to the copy instead of the
original, so that the second application of the predicates is not
affected by the first application.

This is the pattern that is used most other places where paths are
proposed first without materialization of predicates and then with
materialization. (The exceptions are ProposeAccessPathForBaseTable()
and ProposeAccessPathForIndex(), but they reset num_output_rows to
num_output_rows_before_filter before reapplying the filters, so they
are not double-counting the selectivity anyway.)

Change-Id: Ie43a45fdc329bad113631f96ef2a347ffe973e0d