Bug #33600109 Segmentation fault at ha_innobase::index_type， Improper code handling
Problem:
The server may terminates abruptly because nullptr check is
missing for index instance.  The index object is returned by
innobase_get_index() and the return value may be 'nullptr'.
This means that the input parameter of the function
dict_index_is_spatial() is nullptr. That will cause the server
either hits a debug assert or reports a segmentation fault.

Fix:
Added a nullptr check for index instance.

RB: 27517
Reviewed by : Debarun Banerjee <debarun.banerjee@oracle.com>