Bug#33752983  NDB API sendThread failed to raise its priority
TheReceiverThread-id was hard coded in TransporterFacade::raise_thread_prio(),
such that it tried to raise the receiver thread priority, even when called
from theSendThread.

Patch change ::raise_thread_prio() to take an argument, specifying which thread
to raise the priority for. When called from theSendThread, 'theSendThread'
is specified as an argument.

Change-Id: I195e1ae3a9f2bf0df8a1f08fc4a5e34c00842e14