Bug#33670002 SPJ: 'Got error 4829 'FirstInner/Upper has to be an ancestor or a sibling'
The offending query in the test case has the join nest structure:

    t2, t1, (t3, (t5, t4))

Where neither the join conditions on t5 nor t4 has any references
(or explicit 'dependencies') on the 'upper' table t3.
However, they *are* dependent on t3 by being in a nest inside the
same nest as t3. (an implicit 'nest dependency')

When preparing a pushed join, ha_ndbcluster.cc maintain all required
table dependencies between tables and join-nests by adding them to the
'm_ancestor' bitmask of each table.

For the nest level dependencies, they should all be added to the 'first'
table in the relevant nest.

When the relevant dependencies for a specific table is calculated, it
is the sum of all tables being explicitly refered in the join condition,
pluss the implicit dependencies due to the join nests it is a member of -
Limited by the uppermost table it refers in the join condition.

For this particular join query we did not properly take into account
that there might not be any referenced to tables in the closest 'upper'
nest at all (the nest starting with t3).

In such cases we are dependent on all nests up to we reach the nest
containg the uppermost table being referred.

Patch fix this by introducing a while-loop where we add ancestor
nest dependencies until we reach this uppermost tables.
(In the common cases we will loop at most 1 time over this loop.)

Change-Id: I44d34669cbd566d07b1211585c7f1ef87a75c723