WL#14672: Enable the hypergraph optimizer for UPDATE [6/8, multi-table]
Enable use of the hypergraph optimizer for multi-table UPDATE
statements. Single-table UPDATE will be enabled in a later patch.

Currently, only nested loop joins are used. No hash joins. The old
optimizer has the same limitation, so the executor is not ready for
UPDATE statements using hash joins just yet. Hash joins will be
enabled in a later patch.

The main difference from the old optimizer is that the hypergraph
optimizer keeps track of which plans enable use of immediate update,
in which a table can be updated while it is scanned, instead of
delayed updates after the join is done. It keeps alternative subplans
that enable immediate update even if their cost is higher than a
subplan that does not enable immediate update, and makes the final
decision only when the full plan is constructed and the cost of
buffering row IDs in temporary tables has been taken into account.

The old optimizer, on the other hand, only looks at the cost of the
join, and only enables immediate update if the cheapest join plan
happened to make it possible to enable it. It does not consider using
a slightly costlier join in order to avoid the cost of buffering row
IDs for delayed updates.

Change-Id: I1a1e7657d83b0b90750a1fcdb6100cffb6f0760b