Bug#33769911 With ROLLUP, projection size may be needlessly large, twice in select list
In a grouped query with rollup and a window function which orders on an
aggregate which is already present in the SELECT list, an extra
aggregate is added as a hidden column when resolving the window's
ordering clause. This happens due to the fact that matching fails when
the existing SUM(j) has been wrapped in a Item_rollup_sum_switcher (if
WITH ROLLUP is not present, the match will succeed and there is only
one instance of SUM(j) in the select list after resolution).

Fix: extend Item_sum::eq to "see through" Item_rollup_sum_switcher so
that a match will be found.

Since this is not a functional change (only performance is impacted),
no new tests are added. Verified by using a debugger.

Change-Id: I9d4dd9111ea9ad4849b26c32819f32e364277348