Bug#33669039 SPJ read stale row, undefined result, hit assert(key_cmp_if_same..
We hit an assert in ha_ndbcluster handler interface when we fetch a
result row from a table being a member in a pushed join. That assert
checks that the fetched row has a key value according to the key
specified as an argument to the handler method. Note that for rows
returned from pushed joins the 'keys' are not used to locate the result
rows at all - Instead we use a 'correlation Id' which correlate rows
from child tables with its parent table. However the key from the
returned rows should still match though.

The root cause seems to be that the NdbQuery-Interface (aka SPJ API)
is not fully prepared for the Iterator-execution introduced in 8.0.x
The SPJ API has a requirement of results from all pushed tables being retrieved
'in order'. Such that before fetching a row from NdbQueryOperation 'n',
we need to establish a new 'current row' from all NdbQueryOperations
preceeding it (... which it depends on).

For the particular query in this bug report, it has the join-nest
structure: t1 oj (t2,t3) oj (t4,t5). In addition there are join
condition dependencies between the (t4,t5) nests, and t3
in the (t2,t3) outer-nest.

When executed by the iterator-executor it might fail to find
t1-t2 match, which also conclude that the (inner-join) nest
(t2,t3) is a non-match. The executor might then skip directly
to find matched for the nest (t4,t5) - *without* ever fetching
a row for t3!, Thus, the 'NdbResultStream' for t3 will still
refer any previous result rows from t3 as the 'current row' when
we start fetching t4 rows. (by correlation Id's)

The t4-key refer column values from t3, and thus fails the assert
when comparing the key against the stale values.

Patch introduce a bitmask of the NdbQueryOperation having a valid
'currentRow'. When a new first- or next-row has been fetched, that
row becomes valid, but also invalidates rows from all other tables
which depended upon the previous value of that row.

This bitmask is used together with checks for m_currentRow when
we navigate the pushed join result set, such that a (previously)
valid m_currentRow is considdered non-existing if the 'valid-bit'
is not set as well.

Change-Id: I69aebe60b544b06056fae5fa454b7bb9f348c9c3