Bug#33783762 routertest_component_config fails with Address already in use
Issue
=====

In routertest_component_config the test

UnknownConfigOptionWarningCaseInsensitive/
  UnknownConfigOptionWarningCaseInsensitiveTest.UnknownConfigOptionWarningCaseInsensitive/5

sometimes fails with:

  failed to bind(127.0.0.1:11070): Address already in use

The test allocates the port at test-create-time, but falsely releases it
before the test is done, allowing another parallel running test to allocate
the same port.

Change
======

Allocate the tcp-port when the test is actually run and keep it
allocated until the test is finished.

Change-Id: I8c98f43c3674cba91ae1bfb117ce5652f449efab