Bug #33657235 Assertion failure: lock0lock.cc:NNN:!other_lock thread NNN
The code wrongly assumed, that if an indexed column has changed (according to `row_upd_changes_ord_field_binary(..)`) then it must mean that a new record for it should be created in the secondary index.
Unfortunatelly, in case of utf8mb4 strings, the `row_upd_changes_ord_field_binary(..)` method may report a false positive, because it compares `N*4` bytes, as opposed to `N` chars.
So, for example, if an UPDATE changes the column value from "ee" to "ef", and the index is on just first character, then such change doesn't really require any modification to the secondary index, yet the function reports that it might have changed.
One way to solve this would be to fix this function to use the proper encoding for comparison, but it seems to be implemented this way by design (see "binary" in its name).
It seems simpler and more robust for me, to simply use the function which properly converts implicit to explicit lock only when the implicit lock is really held: `lock_rec_convert_impl_to_expl`.
This way, no matter what other special cases we forgot about, it should just work.

RB: 27506
Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
Change-Id: Ia00a5a6564159dbd17a789b60dee30d409dac230