Bug#33760236: ClusterJ test fails when cross compiled in Java 11 for Java 8
The MTR clusterj test fails if the clusterj jar was compiled using
Java 11 and the tests are run using a lower Java version. The
compilation target is already set appropriately but due to a missing
argument 'bootclasspath', the compiler used the latest API, which was
Java 11 API in this case, to compile the code and thus causing issues
when the tests ran in lower Java versions. This patch fixes that by
replacing the target, source javac arguments with the newer 'release'
argument for Java versions 9 and above.

Change-Id: I7cf2888d0a2e9e0bcf8d126d385ce60f74d1ddb7