Bug#31884434: ASSERTION "HA_INNOBASE..TRY_SEMI_CONSISTENT_READ"
EXPLAIN of multi-table UPDATE statements that contained derived tables
sometimes failed and made the server exit.

The problem was caused by Query_result_update::cleanup(), which
disables semi-consistent reads on the outermost table in the join
order after the statement is done. If the outermost table is a
temporary table for a derived table, it does not exist (because the
temporary table is created by MaterializeIterator::Init(), and EXPLAIN
does not call Init() on the iterators). And disabling semi-consistent
reads on a non-existing table behaves badly in some engines.

Fixed by not attempting to disable semi-consistent reads on
non-existing tables during cleanup.

Change-Id: Ifebdc69bc1863c2d08fe01b1502239df97c5a0bc