Bug#33788832: wrong result with COALESCE function
The problem is with certain BETWEEN operations, when the arguments
to BETWEEN are of type DATETIME and the operation is performed using
a time type (not character string). When the second or third argument
to BETWEEN is NULL, the comparison results are inverted, which causes
the function to return NULL for FALSE and FALSE for NULL.

The fix to the problem is to invert the comparison.

Change-Id: I7fbc7ae848f2a14e4dbc58ba26661b6eeb281f18