Bug#33787914: int-decimal. using left join, the returned result is incorrect
When writing DECIMAL values to the hash join buffer, the values are
normalized first. This normalization may in some cases unintentionally
alter the values, which makes the values fall into the wrong hash
bucket and causes wrong results to be returned from the join.

The problem is that the normalization removes leading zeros from the
my_decimal object by changing the number of integer digits (the "intg"
member), without adjusting the "buf" member accordingly. If the whole
first word of "buf" is zero, this incorrect transformation loses some
of the least significant digits.

This normalization is not strictly necessary. It is done in order to
ensure that we write the exact same bytes for two equal decimal
values, regardless of the number of leading zeros the values have. But
we don't need to change the representation in order to achieve that;
it is enough that we pass the same precision and scale arguments to
my_decimal2binary().

Fixed by calculating the desired precision and passing the calculated
value to my_decimal2binary() instead of changing the my_decimal value.

Change-Id: Ie2ce5ce7a01780e3dd3cdeced0bde57ad393e3f2