Bug#33783274 EventBuffer hash-key generation ignore most signifficant byte of key
When the NdbEventBuffer API generate a hash-key for any non-character
data type it use the 'my_charset_bin' hash method, which effectively
ends up calling the my_hash_sort_bin() function.

The 'hash_sort' methods takes two initial seed arguments, the 'nr1' and
'nr2', which are used to fold the binary data into a hash key. For each byte
being hashed, the byte is multiplied with a weight-factor calculated from
these two seeds as:

  hash_weight = ((((uint)nr1 & 63) + nr2

If both seeds are '0', the hash_weight will become '0' as well!
Thus the hash contribution from the first byte in the hash-key is always
multiplied by '0', and do not contribute to the uniqueness of the hash-key.

In the likely case where the first column in the key is an int-type,
represented as a little-endian, this will result in that all numbers 0..255,
256..511 and so forth will all generate the same hash key.

Patch change the intial seed for 'nr2' from '0' to '4' (as used in
most other places where we call 'hash_sort'

It also add handling of zero-length strings, such that they contribute
to the hash key as well, instead of just being ignored.

Change-Id: Ie298443dac8d30c990536fe7b3ba4e0098362252
(cherry picked from commit 34b4f4e792ab018ff71222a6fa66ebd085a74b23)